Ryans requests
